# Image transformation chain reimplementation