TL;DR: I found a way to optimize resizing of JPEG-images significantly. Many changes had to be done in order to get this to work.
## Background

I was benchmarking Imagick against a different image backend and turned to Imbo to see how we had implemented the `jpeg:size`-flag, and was surprised to find that it didn't actually give us any speed increase. After a few minutes, I quickly realized that this is caused by us not setting the flag prior to loading the image blob into Imagick.

Obviously I wanted to fix this in Imbo, but it turned out to be quite a lot of work:
- I needed a way to get the size of the smallest input image the transformation chain could accept. This is already done in the image variations listener, but I never liked how it has transformation names hardcoded into it, making it impossible for third-party resize transformations to hook onto this.
- There was no simple way to get the class associated with a given transformation name. You had to simply trigger an event and see if anything picked it up. To make things worse, if multiple transformations of the same name was in the same chain, it was impossible to see which one returned the smallest size.
- Some transformations require the full image, and some transformations alter the image in a way that makes it hard or impossible for subsequent transformations to return the correct minimum input size. For instance, rotating an image 37 degrees and applying a `maxSize`-transformation after this is not resizing just the actual image, but instead a rather weird canvas.
## Changes

To try and summarize the changes done in this branch:
- Imbo now has a `TransformationManager`, which keeps track of all the available transformations. Transformations are no longer configured in the `eventListeners` configuration section, instead they are now in a separate `transformations` section.
- The manager is what triggers the transformations, and should in theory be able to pass on more information to each transformation if we want it to. For instance, some transformations might want to know if the image has been rotated or resized in order to correctly calculate where to draw objects (`drawPois` being one of them)
- Two new interfaces have been added that each transformation can implement:
  - `InputSizeConstraint` - will signal to the manager that this transformation needs a certain input size. Typically this is for transformations that do resizing - `maxSize`, `thumbnail` and `resize` all implement it. It can also be used to indicate that it needs the full size image, that the result of the transformation is unpredictable and should stop the minimum size resolving and can also tell Imbo if any rotation was performed. It also contains a method (`adjustParameters()`) that transformations can implement to adjust any parameters that they use, in case the input size was changed. Currently this is used for `canvas`, `border`, `watermark` and `crop`, which all use coordinates or width/height parameters.
  - `RegionExtractor` - will signal to the manager that the transformation extracts a portion of the image. This is used in combination with the minimum input size resolving in order to find the smallest possible size. `crop` currently implements this.
- Imagick asks the transformation manager what the minimum input size is and sets the `jpeg:size` hint before loading the image blob. This allows libjpeg to use a "shrink-on-load" implementation, which significantly increases the speed of the resizing.
- The image variations listener uses the transformation manager in order to resolve minimum variation size and to adjust transformation parameters. This decouples it from core and enables third-party transformations to benefit from it.
## Result

To benchmark the old implementation against the new, I wrote a script that:
- Queries for 300 images (all 2000+ pixels in width)
- Loops through each image and generates an image url with a `maxSize` transformation set to 523x315
  - Image variations are generated for all the images, closest size to 523px being 1000px
- Downloads the image URLs
- Iterates over the same image set 3 times to remove any outliers

This is all done against a server running locally, so network conditions shouldn't have any say. I've reproduced the results multiple times with just small deviations. The result of resizing 900 images (300 \* 3 iterations):

| Implementation | Variations enabled | Speed |
| --- | --- | --- |
| Old | No | 123.01s |
| New | No | 45.57s |
| Old | Yes | 22.30s |
| New | Yes | 16.73s |
## Conclusions

I think the results speak for themselves. Most importantly, this is a "free" optimization - it doesn't require any more disk space or configuration to set up. Obviously, this only applies to JPEG-images that is able to use the "shrink-on-load" trick, but I'm guessing most peoples image collection consists primarily of JPEGs anyway.
## Further work

First and foremost, the `smartSize` transformation has not been optimized. I think @kbrabrand should do this, being the author of it, and it might be that the transformation manager will require some small changes for it to work with this transformation. 

We should also give this some more testing in a production-like environment. Many combined transformations and order of these might provoke bugs, so we should also try quite a bit of transformation combinations before merging.