@kbrabrand 2015-12-07
This is awesome! :clap: I've added a comment with a link to a gist in #393. Will try to find time to look at getting this optimization to work with `smartSize` soon.