@christeredvartsen 2015-12-07
Really nice work @rexxars! The way the transformations are handled here enables more interaction between the transformations which is a nice addition.