@matslindh 2016-01-13
Since `thumbnail` in the original imagemagick code uses a 5x for this operator, so we might want larger than 2x as a default value (my vote would be for 3x, since that would match a 3x3 core around the pixel to be resized, which usually give a good result with bilinear resizing). Just using the hint should be enough, I'm fairly sure imagick will pick the most suitable of the original size and the `jpeg:size` setting anyway.

I'm just not sure if we really need to make this configurable at all right now - another setting to maintain and keep supporting in the future.