@christeredvartsen 2016-03-16
@rexxars Care to do a rebase against develop?

@kbrabrand Could you see if you might find some time to implement support for the `smartSize` transformation as well?