@kbrabrand 2016-03-17
I'll try, but it's been really busy lately with courses at university and the new job :/